% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articlecount.R
\name{get_article_vc}
\alias{get_article_vc}
\title{Return a period page view count of a specific article from wikipedia page view api.
This function generate a data frame contains a specific article page view count within an entered period.}
\usage{
get_article_vc(
  article_title,
  starting = paste(substr(toString(Sys.Date() - 2), 1, 4), substr(toString(Sys.Date() -
    2), 6, 7), substr(toString(Sys.Date() - 2), 9, 10), sep = ""),
  ending = paste(substr(toString(Sys.Date() - 1), 1, 4), substr(toString(Sys.Date() -
    1), 6, 7), substr(toString(Sys.Date() - 1), 9, 10), sep = ""),
  period = "daily"
)
}
\arguments{
\item{article_title}{the article name, selected from wikipedia For example: "https://en.wikipedia.org/wiki/Linear_algebra" article_title is "Linear_algebra"}

\item{starting}{period start timestamp with format YYYYMMDD, for example 20200102(Jan 2nd, 2021), with the default value the timestamp of the day before yesterday}

\item{ending}{period start timestamp with format YYYYMMDD, for example 20200102(Jan 2nd, 2021), with the default value the timestamp of yesterday}

\item{period}{period type should be 'daily' or 'monthly', with the default value 'daily'}
}
\value{
return a 2-column data frame
}
\description{
Return a period page view count of a specific article from wikipedia page view api.
This function generate a data frame contains a specific article page view count within an entered period.
}
\examples{
get_article_vc("Linear_algebra", "20150803", "20201002", "monthly")
}
