% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projtop.R
\name{get_project_top}
\alias{get_project_top}
\title{Return a period page view count of a specific project top ranking from wikipedia.
This function generate a data frame contains a specific project page view count within an entered period.}
\usage{
get_project_top(
  project_title,
  top = "top",
  tool = "all-access",
  year = substr(toString(Sys.Date() - 2), 1, 4),
  mon = substr(toString(Sys.Date() - 2), 6, 7),
  day = substr(toString(Sys.Date() - 2), 9, 10)
)
}
\arguments{
\item{project_title}{the project name, values from ISO language code(ISO 639-1) plus project type(e.g. en.wikipedia, ca.wikinews, zh.wikisource)
project_title refer to https://en.wikipedia.org/wiki/Main_Page (In other projects;languages)}

\item{top}{#values in ("top", "top-by-country"), with default "top"}

\item{tool}{access tool,values in ("all-access", "desktop", "mobile-app", "mobile-web") are allowed for tool, with default "all-access"}

\item{year}{year time with format YYYY, for example 2020, with the default value the year of the day before yesterday}

\item{mon}{month time with format MM, for example 01(January), with the default value the month of the day before yesterday}

\item{day}{day time with format DD, for example 03(third), with the default value the day before yesterday}
}
\description{
Return a period page view count of a specific project top ranking from wikipedia.
This function generate a data frame contains a specific project page view count within an entered period.
}
\examples{
get_project_top("ja.wikisource", top = "top", year = "2018", mon = "12", day = "12")
}
