% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projud.R
\name{get_project_ud}
\alias{get_project_ud}
\title{Return a project unique device count of a specific period for all sites.
This function generate a data frame contains a projet unique device count within an entered period.}
\usage{
get_project_ud(
  project_title,
  method = "all-sites",
  starting = paste(substr(toString(Sys.Date() - 2), 1, 4), substr(toString(Sys.Date() -
    2), 6, 7), substr(toString(Sys.Date() - 2), 9, 10), sep = ""),
  ending = paste(substr(toString(Sys.Date() - 1), 1, 4), substr(toString(Sys.Date() -
    1), 6, 7), substr(toString(Sys.Date() - 1), 9, 10), sep = ""),
  period = "daily"
)
}
\arguments{
\item{project_title}{the project name, values from ISO language code(ISO 639-1) plus project type(e.g. en.wikipedia, ca.wikinews, zh.wikisource)
project_title refer to https://en.wikipedia.org/wiki/Main_Page (In other projects;languages)}

\item{method}{method used to view page, values in ("all-sites", "desktop-site", "mobile-site") are allowed for method, with default "all-sites"}

\item{starting}{period start timestamp with format YYYYMMDD, for example 20200102(Jan 2nd, 2021), with the default value the timestamp of the day before yesterday}

\item{ending}{period start timestamp with format YYYYMMDD, for example 20200102(Jan 2nd, 2021), with the default value the timestamp of yesterday}

\item{period}{period type should be 'daily' or 'monthly', with the default value 'daily'}
}
\description{
Return a project unique device count of a specific period for all sites.
This function generate a data frame contains a projet unique device count within an entered period.
}
\examples{
get_project_ud("zh.wikipedia", method = "desktop-site", start = "20161010", end = "20171012")
}
