% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projcount.R
\name{get_project_vc}
\alias{get_project_vc}
\title{Return a period page view count of a specific project from wikipedia page view api.
This function generate a data frame contains a specific project page view count within an entered period.}
\usage{
get_project_vc(
  project_title,
  tool = "all-access",
  method = "all-agents",
  starting = paste(substr(toString(Sys.Date() - 2), 1, 4), substr(toString(Sys.Date() -
    2), 6, 7), substr(toString(Sys.Date() - 2), 9, 10), substr(toString(Sys.time() - 2),
    12, 13), sep = ""),
  ending = paste(substr(toString(Sys.Date() - 1), 1, 4), substr(toString(Sys.Date() -
    1), 6, 7), substr(toString(Sys.Date() - 1), 9, 10), substr(toString(Sys.time() - 1),
    12, 13), sep = ""),
  period = "daily"
)
}
\arguments{
\item{project_title}{the project name, values from ISO language code(ISO 639-1) plus project type(e.g. en.wikipedia, ca.wikinews, zh.wikisource)
project_title refer to https://en.wikipedia.org/wiki/Main_Page (In other projects;languages)}

\item{tool}{access tool,values in ("all-access", "desktop", "mobile-app", "mobile-web") are allowed for tool, with default "all-access"}

\item{method}{method used to view page, values in ("all-agents", "user", "spider", "automated") are allowed for method, with default "all-agents"}

\item{starting}{period start timestamp with format YYYYMMDD, for example 20200102(Jan 2nd, 2021), with the default value the timestamp of the day before yesterday}

\item{ending}{period start timestamp with format YYYYMMDD, for example 20200102(Jan 2nd, 2021), with the default value the timestamp of yesterday}

\item{period}{period type should be 'daily' or 'monthly', with the default value 'daily'}
}
\description{
Return a period page view count of a specific project from wikipedia page view api.
This function generate a data frame contains a specific project page view count within an entered period.
}
\examples{
get_project_vc("zh.wikipedia", method = "user", starting = "2021010112", period = "hourly")
}
