% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcount.R
\name{plotcount}
\alias{plotcount}
\title{Return a plot of a count data frame.
This function generate a plot by data from the entered data frame.}
\usage{
plotcount(viewcount)
}
\arguments{
\item{viewcount}{data frame entered for plotting}
}
\value{
return a message when try plot top ranking plot, or a plot for other data frame
}
\description{
Return a plot of a count data frame.
This function generate a plot by data from the entered data frame.
}
\examples{
plt1 <- get_project_top("ja.wikisource", top = "top", year = "2018", mon = "12", day = "12")
plt2 <- get_project_ud("zh.wikipedia", method = "desktop-site", start = "20161010", end = "20171012")
plotcount(plt1)
plotcount(plt2)
}
